Rename transition_run_thread.py to transition_run.py
